Update a database migration

+++ db/migrate/20240717092526_remove_thread_id_from_messages.rb
@@ -1,5 +1,7 @@
 class RemoveThreadIdFromMessages < ActiveRecord::Migration[7.1]
   def change
-    remove_column :messages, :thread_id, :integer
+    if column_exists?(:messages, :thread_id)
+      remove_column :messages, :thread_id, :integer
+    end
   end
 end
